requirements.txt
pytest~=6.2.5
PyYAML~=5.4.1
eth-brownie~=1.18.1